refactor: move mail types into src/types
