+++ src/types/index.ts
@@ -7,6 +7,7 @@ export * from './form_logo'
 export * from './form_statistics_total'
 export * from './form'
 export * from './login'
+export * from './mail'
 export * from './myinfo_hash'
 export * from './response'
 export * from './sms_count'
